refactored object locations
